* appveyor-check-test-suite:
  Appveyor testing expanded
  Fix array overrun in li_carrays testcase
  Warning fixes in generated Python code for 64bit Visual C++ on Windows.
  Warning fixes in generated C# code for 64bit Visual C++ on Windows.
  Warning fixes for 64bit visual c++ on Windows
  Warning fixes in generated Java code for 64bit Visual C++ on Windows.
  Warning fixes for 64bit visual c++ on Windows
  C# gc tests failure fix
  Add a space between literal and string macro

+++ Source/DOH/base.c
@@ -146,7 +146,7 @@ int DohLen(const DOH *obj) {
     }
     return 0;
   } else {
-    return strlen((char *) obj);
+    return (int)strlen((char *) obj);
   }
 }
 
@@ -636,7 +636,7 @@ int DohRead(DOH *obj, void *buffer, int length) {
     return -1;
   }
   /* Hmmm.  Not a file.  Maybe it's a real FILE */
-  return fread(buffer, 1, length, (FILE *) b);
+  return (int)fread(buffer, 1, length, (FILE *) b);
 }
 
 /* -----------------------------------------------------------------------------
@@ -654,7 +654,7 @@ int DohWrite(DOH *obj, const void *buffer, int length) {
     return -1;
   }
   /* Hmmm.  Not a file.  Maybe it's a real FILE */
-  return fwrite(buffer, 1, length, (FILE *) b);
+  return (int)fwrite(buffer, 1, length, (FILE *) b);
 }
 
 /* -----------------------------------------------------------------------------
